requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
python-dotenv==1.0.0
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.3